# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API",
        "contact": {
            "name": "Support Team",
            "email": "support@my-api.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/color": {
            "get": {
                "tags": [
                    "Color"
                ],
                "summary": "Obtener lista de colores",
                "description": "Devuelve una lista paginada de colores",
                "operationId": "c25bb73230a25c40e0ea6f49be0a9632",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Número de colores por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "Descripcion": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Color"
                ],
                "summary": "Crear un nuevo color",
                "description": "Crea un nuevo color y lo devuelve",
                "operationId": "2853e11ec540dd436db1a02f95e05211",
                "requestBody": {
                    "description": "Color a crear",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "Descripcion": {
                                        "description": "La descripción del color",
                                        "type": "string",
                                        "maxLength": 255
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Datos de entrada no válidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/color/{id}": {
            "get": {
                "tags": [
                    "Color"
                ],
                "summary": "Obtener un color",
                "description": "Devuelve un color específico",
                "operationId": "82d00db326bce07adfeb198fd58bc3bd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del color a obtener",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Descripcion": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Color no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Color"
                ],
                "summary": "Actualizar un color",
                "description": "Actualiza un color existente y lo devuelve",
                "operationId": "a051c9898924a9ed961ac0eb24f2ef95",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del color a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Color a actualizar",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "Descripcion": {
                                        "description": "La descripción del color",
                                        "type": "string",
                                        "maxLength": 255
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Datos de entrada no válidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Color no encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Color"
                ],
                "summary": "Eliminar un color",
                "description": "Elimina un color existente",
                "operationId": "f50e6c1b70b1a2550b6cd496c28df0dd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del color a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Color no encontrado"
                    }
                }
            }
        },
...
    },
...
}